Use retry function instead of hacky try/catch
Not only is Jenkins' built-in retry() function much less verbose than
the previous try/catch block, but it will correctly throw an exception
should the container fail to start.

The previous implementation could have been fixed to do this as well,
by checking if the retry count was zero or negative and then throwing
an exception after the while() loop, but this is a much better
solution.

The README has been updated accordingly, since failures in a `retry`
closure are not shown with scary red X's in Blue Ocean.