Annotate getRandomDigitString() with @NonCPS
I suspect that the IntRange created by the `(0..length - 1).collect`
statement is causing java.io.NotSerializableException to be thrown if
Jenkins attempts to save build state when this method is being
executed. Unfortunately, it's a bit hard to verify with a high degree
of certainty that this is the cause of such exceptions, but
fortunately we are not calling any pipeline steps from this method so
it is safe to annotate with @NonCPS anyways. However, IntRange (which
is created by using the `..` operator) does not implement the
Serializable interface, so it is reasonable to assume that this could
cause problems.

Unfortunately, we can't directly provide a test for this, but the
annotation should prevent Jenkins from attempting to serialize state
while in this method.

For more information about serialization in pipelines and @NonCPS,
see: https://github.com/jenkinsci/pipeline-plugin/blob/master/TUTORIAL.md#serializing-local-variables