Don't create a special user for container
As noted in the Postgres Docker documentation (in the "Arbitrary
--user Notes" section), Postgres doesn't care what user it is running
as, but we must add user and group mapping arguments to `dockerArgs`
to ensure that the permissions for the data directory are correct. We
must also add a read-only volume mapping for /etc/passwd to fix the
'The files belonging to this database system will be owned by user
"postgres"' error. Note that these flags must be passed to both the
database container and the child container which checks when the
database is ready.

Creating a user with `useradd` as was done before can cause this
library to fail if the UID is the same as any user which already
exists in the container. This is most problematic with UID 999, which
exists on many Linux systems but also is used for the `postgres` user
in the container itself.

Since the custom user was the only reason that we needed to create a
Dockerfile, this commit also eliminates that, which simplifies the
code in this class quite a bit.